requirements_dev.txt
anonapi=1.0.8
pip==19.2.3
bump2version==0.5.11
factory-boy==2.12
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0
Click==7.0
pytest==4.6.5
pytest-runner==5.1
black==19.10b0